+++ packages/core/lib/utils/store/store.dart
@@ -1,4 +1,4 @@
-import 'package:analytics/utils/store/impl.dart';
+import 'package:segment_analytics/utils/store/impl.dart';
 
 mixin Store {
   Future<Map<String, dynamic>?> getPersisted(String key);