+++ packages/plugins/plugin_advertising_id/CHANGELOG.md
@@ -1,3 +1,3 @@
-## 0.0.1
+## 1.0.0
 
-* Initial version, based off the React Native implementation (@segment/analytics-react-native-plugin-advertising-id) v1.2.0
+- Initial version, based off the React Native implementation (@segment/analytics-react-native-plugin-advertising-id) v1.2.0