+++ packages/plugins/plugin_idfa/CHANGELOG.md
@@ -1,3 +1,3 @@
-## 0.0.1
+## 1.0.0
 
-*  Initial version, based off the React Native implementation (@segment/analytics-react-native-plugin-idfa) v0.5.0
+- Initial version, based off the React Native implementation (@segment/analytics-react-native-plugin-idfa) v0.5.0